E1/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.24.3
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
plotly==5.17.0
python-multipart==0.0.6
chardet==5.2.0 
python-jose[cryptography] 
passlib[bcrypt] 
python-multipart
pydantic[email]